==> 02_feature_engineering.ipynb <==
# data manipulation
import numpy as np
import pandas as pd

# from feature-engine
from feature_engine.encoding import (
    OrdinalEncoder,
    OneHotEncoder,
)

# from Scikit-learn
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler

# Pickle for data preprocessing
import pickle

df= pd.read_csv(r"C:\Users\furka\JupyterNotebookProjects\End-to-End-Churn-Modelling-Binary-Classification\Churn_Modelling.csv")

df.head()

## Drop the unnecessary columns
df=df.drop(['RowNumber','CustomerId','Surname'],axis=1)
df

# Drop the null values.
# Since the number of missing data is low, we can directly delete the missing data instead of filling it in. 
# Because it will not affect the model. We can use dropna() for this.
df = df.dropna()
# Dropping the duplicate values
df = df.drop_duplicates()

print('Rows containing missing data and duplicate data were deleted.')

df

# Separate into train and test set
# Remember to set the seed (random_state for this sklearn function)
X_train, X_test, y_train, y_test = train_test_split(
    df.drop(['Exited'], axis=1), # predictive variables
    df['Exited'], # target
    test_size=0.2, # portion of dataset to allocate to test set
    random_state=0, # we are setting the seed here
)

print('The dataset separate dataset into train and test successfully')
X_train.shape, X_test.shape

- Label Encoding for 'Gender' column

label_encoder = OrdinalEncoder(encoding_method='arbitrary', variables='Gender')

# fit the encoder
label_encoder.fit(X_train)

# mappings are stored and class can be saved
label_encoder.encoder_dict_

X_train = label_encoder.transform(X_train)
X_test = label_encoder.transform(X_test)

X_train

X_test

- One Hot Encoding for 'Geography' column

# set up encoder
oh_encoder = OneHotEncoder(
    variables='Geography',
    drop_last=False,  # to return k-1, use drop=false to return k dummies
)

# fit the encoder
oh_encoder.fit(X_train)

# mappings are stored and class can be saved
oh_encoder.encoder_dict_

X_train = oh_encoder.transform(X_train)
X_test = oh_encoder.transform(X_test)

X_train

X_test

# Save the encoders
with open('label_encoder_gender.pkl','wb') as file:
    pickle.dump(label_encoder,file)

with open('oh_encoder_geo.pkl','wb') as file:
    pickle.dump(oh_encoder,file)

X_train.head()

X_test.head()

- Feature Scaling

# create scaler
scaler = StandardScaler()

#  fit  the scaler to the train set
scaler.fit(X_train) 

# transform the train and test set

# sklearn returns numpy arrays, so we wrap the
# array with a pandas dataframe

X_train = pd.DataFrame(
    scaler.transform(X_train),
    columns=X_train.columns
)

X_test = pd.DataFrame(
    scaler.transform(X_test),
    columns=X_train.columns
)

X_train.head()

X_test.head()

with open('scaler.pkl','wb') as file:
    pickle.dump(scaler,file)

- All Preprocessing and feature engineering is done
