==> 05_predict.ipynb <==
import tensorflow as tf
from tensorflow.keras.models import load_model
import pickle
import pandas as pd
import numpy as np

# Load the trained model and preprocessing pkl files
model=load_model('ann_model.h5')

## load the encoder and scaler
with open('oh_encoder_geo.pkl','rb') as file:
    oh_encoder_geo=pickle.load(file)

with open('label_encoder_gender.pkl', 'rb') as file:
    label_encoder_gender = pickle.load(file)

with open('scaler.pkl', 'rb') as file:
    scaler = pickle.load(file)

oh_encoder_geo.encoder_dict_['Geography']

label_encoder_gender.encoder_dict_['Gender']

# Example input data
input_data = {
    'CreditScore': 600,
    'Geography': 'France',
    'Gender': 'Male',
    'Age': 40,
    'Tenure': 3,
    'Balance': 60000,
    'NumOfProducts': 2,
    'HasCrCard': 1,
    'IsActiveMember': 1,
    'EstimatedSalary': 50000
}

input_df = pd.DataFrame([input_data])

input_df = label_encoder_gender.transform(input_df)

input_df

input_df = oh_encoder_geo.transform(input_df)

input_df

input_df = pd.DataFrame(
    scaler.transform(input_df),
    columns=input_df.columns
)

input_df

## Predict churn
prediction=model.predict(input_df)
prediction

prediction_proba = prediction[0][0]
prediction_proba

if prediction_proba > 0.5:
    print('The customer is likely to churn.')
else:
    print('The customer is not likely to churn.')
